+++ examples/kong-gateway-operator/setup.sh
@@ -5,20 +5,10 @@ OPENBAO_EXISTS=$(kubectl get ns openbao || echo "false")
 
 if [ "$OPENBAO_EXISTS" == "false" ]
 then
-cd ../openbao/
+cd ../openbao/ || exit 1
 bash setup.sh
 else
 echo "Skipping OpenBao deployment. Already there."
-fi
-
-# Remove HAProxy Ingress - will be replaced with Kong Gateway resp. an Operator-based Gateway
-HAPROXY_EXISTS=$(kubectl get ns ingress-haproxy 2>/dev/null || echo "false")
-if [ "$HAPROXY_EXISTS" == "false" ]
-then
-echo "Skipping deletion of HAProxy ingress..."
-else
-kubectl delete -f ../../manifests/haproxy-helm.yaml || true
-kubectl delete ingress -n demo httpbin
 fi
 
 echo "\nInstall Gateway API extension"
@@ -46,7 +36,7 @@ kubectl rollout restart deployment cert-manager -n cert-manager
 
 kubectl -n cert-manager wait --for=condition=Available=true --timeout=120s deployment/cert-manager
 
-cd ../kong-gateway-operator/
+cd ../kong-gateway-operator/ || exit 1
 
 echo "\nInstall Kong Gateway Operator"
 helm repo add kong https://charts.konghq.com